+++ docs/history.md
@@ -3,8 +3,10 @@
 <!-- cspell:ignore cantgirl,Fallenfor,Finalkids,Hispa,Kaethela,lenamek,mgln,Moucheron,Quipet,Sanio,sisuka,SuperpaperChildy,Toooschi -->
 <!-- markdownlint-disable MD001 -->
 
-### _v1.8.1_ - September 11th, 2023 (Unreleased)
+### _v1.9.0_ - September 11th, 2023 (Unreleased)
 
+- Changed the following babies:
+  - Silly Baby (#516) - Sacrificial Altar effect after 6 hits
 - Fixed the bug where Turtle Dragon Baby would be softlocked under certain conditions. (Thanks Gamonymous)
 - Fixed the bug where Gold Baby would not work properly with certain floors.
 - Fixed the bug where babies would transformations would not be granted under certain conditions. (Thanks Monkeypot)